requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
ezdxf==1.4.3
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
fonttools==4.62.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.3
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyparsing==3.3.2
SQLAlchemy==2.0.48
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.6
